Improvments on callable objects in the di.
