+++ src/App/Traits/AjaxTrait.php
@@ -165,8 +165,8 @@ trait AjaxTrait
     public function cl(string $sClassName)
     {
         $sClassName = trim($sClassName);
-        $xCallableClass = $this->xCallableRegistry->getCallableObject($sClassName);
-        return !$xCallableClass ? null : $xCallableClass->getRegisteredObject();
+        $xCallableObject = $this->xCallableRegistry->getCallableObject($sClassName);
+        return !$xCallableObject ? null : $xCallableObject->getRegisteredObject();
     }
 
     /**
